% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BITFAM_activities.R
\name{BITFAM_activities}
\alias{BITFAM_activities}
\title{Extract the TF inferred activities from the output of BITFAM main function}
\usage{
BITFAM_activities(BITFAM_list)
}
\arguments{
\item{BITFAM_list}{A list generate by BITFAM main function}
}
\value{
TF inferred activities of each cell
}
\description{
Extract the TF inferred activities from the output of BITFAM main function
}
