% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BITFAM_preprocess.R
\name{BITFAM_preprocess}
\alias{BITFAM_preprocess}
\title{Generate the normalized scRNA-seq data by Seurat 3.0}
\usage{
BITFAM_preprocess(raw_data)
}
\arguments{
\item{raw_data}{A matrix or dataframe of scRNA-seq raw counts table}
}
\value{
Normalized scRNA-seq data
}
\description{
Generate the normalized scRNA-seq data by Seurat 3.0
}
