% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BITFAM_scATAC.R
\name{BITFAM_scATAC}
\alias{BITFAM_scATAC}
\title{Extract the genes that have scATAC-seq peaks on their promoter regions}
\usage{
BITFAM_scATAC(scATAC_obj)
}
\arguments{
\item{scATAC_obj}{A preprocessed Seurat object of scATAC-seq data.}
}
\value{
the genes that have scATAC-seq peaks on their promoter regions
}
\description{
Extract the genes that have scATAC-seq peaks on their promoter regions
}
