% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BITFAM_weights.R
\name{BITFAM_weights}
\alias{BITFAM_weights}
\title{Extract the TF-gene weights from the output of BITFAM main function}
\usage{
BITFAM_weights(BITFAM_list)
}
\arguments{
\item{BITFAM_list}{A list generate by BITFAM main function}
}
\value{
TF-gene weights for each pair
}
\description{
Extract the TF-gene weights from the output of BITFAM main function
}
