% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BITFAM.R
\name{BITFAM}
\alias{BITFAM}
\title{BITFAM main function. BITFAM will infer the transcription factor activities from single cell RNA-seq data based on the ChIP-seq data}
\usage{

BITFAM(data, species, interseted_TF = NA, scATAC_obj = NA, iter = 8000, tol_rel_obj = 0.005, ncores)
  
\arguments{
\item{data}{A matrix or dataframe, normalized single cell RNA-seq data}

\item{species}{mouse or human}

\item{interseted_TF}{Transcription factors of interests}

\item{scATAC_obj}{A preprocessed Seurat object of scATAC-seq data}

\item{number}{Number of CPU cores}

\item{iter}{Number of max iteration}

\item{tol_rel_obj}{The convergence tolerance on the relative norm of the objective}
}
\value{
sampling results of TF inferred activities and TF-gene weights
}
\description{
BITFAM main function. BITFAM will infer the transcription factor activities from single cell RNA-seq data based on the ChIP-seq data
}
